moving asset helpers into a separate file
